Add Screen, ESP now, and more
Major update.
We now support one "server" that has a screen connected to it, and connects to a wifi hotspot.
A "client" is also supported, that scans all channels and pairs it self automatically with the "server" using ESP now.

We are missing a way to detect that we have lost connectivity and then reconnect. Will probably just implement a watchdog that reboots the whole thing just to make sure it is alive.